check messages for filtering when they're edited
