+++ lib/app.js
@@ -1,11 +1,11 @@
 const fetch = require('node-fetch')
 const Discord = require('discord.js')
 const heapdump = require('heapdump')
-const { log, sendDM, fetchAuditEntryFor } = require('./common')
+const { log, fetchAuditEntryFor } = require('./common')
 const { updateStream, updateStreams } = require('./twitch')
 const { fetchContributors } = require('./contributors')
 const { fetchBlocked } = require('./blocked')
-const { messageFilter, resetMessageCache, buildUserDetail, buildMessageDetail, deleteMessage, isOurDeletion, pruneDeletedMessages } = require('./security')
+const { proccessFilters, resetMessageCache, buildUserDetail, buildMessageDetail, isOurDeletion, pruneDeletedMessages } = require('./security')
 const { flagMessageIfNeeded, pruneRecentlyFlagged } = require('./flag')
 const config = require('./config')
 const commands = require('./commands')
@@ -84,29 +84,7 @@ client.on('ready', () => {
 })
 
 client.on('message', message => {
-  if (message.author.bot) {
-    return
-  }
-
-  const filteredResult = messageFilter(message, client)
-
-  if (filteredResult) {
-    if (deleteMessage(message)) {
-      return
-    }
-
-    const logs = message.guild.channels.cache.find(c => c.name === config.channels.moderationLogs)
-
-    if (!logs) {
-      return
-    }
-
-    const messageDetail = `message was filtered.\n**Reason:** ${filteredResult}.` + buildMessageDetail(message)
-
-    logs.send(`${buildUserDetail(message.author)}'s ${messageDetail}`)
-      .then(() => sendDM(message.author, `Your ${messageDetail}`))
-      .catch(log.debug)
-
+  if (!proccessFilters(message, client)) {
     return
   }
 
@@ -120,6 +98,10 @@ client.on('message', message => {
   const command = args.shift().toLowerCase()
   log.debug('Received command', command, args)
   commands(message, command, args).catch(log.debug)
+})
+
+client.on('messageUpdate', (oldMessage, newMessage) => {
+  proccessFilters(newMessage, client)
 })
 
 client.on('presenceUpdate', (oldPresence, newPresence) => {
