+++ lib/security.js
@@ -147,6 +147,36 @@ function buildMessageDetail (message) {
   return messageDetail
 }
 
+function proccessFilters (message, client) {
+  if (message.author.bot) {
+    return false
+  }
+
+  const filteredResult = messageFilter(message, client)
+
+  if (!filteredResult) {
+    return true
+  }
+
+  if (deleteMessage(message)) {
+    return false
+  }
+
+  const logs = message.guild.channels.cache.find(c => c.name === config.channels.moderationLogs)
+
+  if (!logs) {
+    return false
+  }
+
+  const messageDetail = `message was filtered.\n**Reason:** ${filteredResult}.` + buildMessageDetail(message)
+
+  logs.send(`${buildUserDetail(message.author)}'s ${messageDetail}`)
+    .then(() => sendDM(message.author, `Your ${messageDetail}`))
+    .catch(log.debug)
+
+  return false
+}
+
 function messageFilter (message, client) {
   if (hasPermissions(message.member) || !message.guild) {
     return false
@@ -298,13 +328,12 @@ module.exports = {
   hasPermissions,
   hasAdminPermissions,
   noPermissions,
-  messageFilter,
+  proccessFilters,
   resetMessageCache,
   assignRole,
   removeRole,
   buildUserDetail,
   buildMessageDetail,
-  deleteMessage,
   isOurDeletion,
   pruneDeletedMessages
 }
